+++ frontend/app/src/pages/main/v1/workers/$worker/index.tsx
@@ -25,7 +25,7 @@ import {
 } from '@/components/v1/ui/dropdown-menu';
 import { useState } from 'react';
 import { RecentWebhookRequests } from '../webhooks/components/recent-webhook-requests';
-import { TaskRunsTable } from '../../workflow-runs-v1/components/workflow-runs-table';
+import { TaskRunsTable } from '../../workflow-runs-v1/components/task-runs-table';
 export const isHealthy = (worker?: Worker) => {
   const reasons = [];
