+++ frontend/app/src/pages/main/v1/workflow-runs-v1/components/task-runs-table.tsx
@@ -1,5 +1,5 @@
 import { DataTable } from '@/components/v1/molecules/data-table/data-table.tsx';
-import { columns } from './v1/workflow-runs-columns';
+import { columns } from './v1/task-runs-columns';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import {
   ColumnFiltersState,
