+++ frontend/app/src/pages/main/v1/workflow-runs-v1/index.tsx
@@ -1,5 +1,5 @@
 import { Separator } from '@/components/v1/ui/separator';
-import { TaskRunsTable } from './components/workflow-runs-table';
+import { TaskRunsTable } from './components/task-runs-table';
 import { TriggerWorkflowForm } from '../workflows/$workflow/components/trigger-workflow-form';
 import { useState } from 'react';
